vk-device(refactor): Consolidate the device and queue objects to simplify the implementation
The device and queue are tightly coupled and the info objects are prerequisites that create this tight coupling.

After multiple attempts to apply some form of modularity, I've decided to simply go with the grain instead of against it.

I may attempt to revisit this in the future, but this is the most pragmatic approach to enforce simplicity for time being.